refactor: Remove unnecessary interface from PeerStore
