+++ network/src/benches/sqlite_peer_store.rs
@@ -20,7 +20,7 @@ fn insert_peer_info_benchmark(c: &mut Criterion) {
             let addr = "/ip4/127.0.0.1".to_multiaddr().unwrap();
             move || {
                 for peer_id in peer_ids.clone() {
-                    peer_store.new_connected_peer(&peer_id, addr.clone(), SessionType::Client);
+                    peer_store.add_connected_peer(&peer_id, addr.clone(), SessionType::Client);
                 }
             }
         })
@@ -32,7 +32,7 @@ fn insert_peer_info_benchmark(c: &mut Criterion) {
             let addr = "/ip4/127.0.0.1".to_multiaddr().unwrap();
             move || {
                 for peer_id in peer_ids.clone() {
-                    peer_store.new_connected_peer(&peer_id, addr.clone(), SessionType::Client);
+                    peer_store.add_connected_peer(&peer_id, addr.clone(), SessionType::Client);
                 }
             }
         })
@@ -46,7 +46,7 @@ fn insert_peer_info_benchmark(c: &mut Criterion) {
             let addr = "/ip4/127.0.0.1".to_multiaddr().unwrap();
             move || {
                 for peer_id in peer_ids.clone() {
-                    peer_store.new_connected_peer(&peer_id, addr.clone(), SessionType::Client);
+                    peer_store.add_connected_peer(&peer_id, addr.clone(), SessionType::Client);
                 }
             }
         })
@@ -61,8 +61,8 @@ fn random_order_benchmark(c: &mut Criterion) {
             let mut peer_store = peer_store.lock();
             for _ in 0..8000 {
                 let peer_id = random_peer_id();
-                peer_store.new_connected_peer(&peer_id, addr.clone(), SessionType::Client);
-                let _ = peer_store.add_discovered_address(&peer_id, addr.clone());
+                peer_store.add_connected_peer(&peer_id, addr.clone(), SessionType::Client);
+                let _ = peer_store.add_discovered_addr(&peer_id, addr.clone());
             }
         }
         c.bench_function("random order 1000 / 8000 peer_info", {
@@ -108,8 +108,8 @@ fn random_order_benchmark(c: &mut Criterion) {
                 let addr = "/ip4/127.0.0.1".to_multiaddr().unwrap();
                 for _ in 0..8000 {
                     let peer_id = random_peer_id();
-                    peer_store.new_connected_peer(&peer_id, addr.clone(), SessionType::Client);
-                    let _ = peer_store.add_discovered_address(&peer_id, addr.clone());
+                    peer_store.add_connected_peer(&peer_id, addr.clone(), SessionType::Client);
+                    let _ = peer_store.add_discovered_addr(&peer_id, addr.clone());
                 }
                 move || {
                     let count = 1000;
