+++ network/src/tests/sqlite_peer_store.rs
@@ -9,11 +9,11 @@ use crate::{
 use std::time::Duration;
 
 #[test]
-fn test_new_connected_peer() {
+fn test_add_connected_peer() {
     let mut peer_store: Box<dyn PeerStore> = Box::new(SqlitePeerStore::temp());
     let peer_id = random_peer_id();
     let addr = "/ip4/127.0.0.1".to_multiaddr().unwrap();
-    peer_store.new_connected_peer(&peer_id, addr, SessionType::Client);
+    peer_store.add_connected_peer(&peer_id, addr, SessionType::Client);
     assert_eq!(
         peer_store.peer_score(&peer_id).unwrap(),
         peer_store.scoring_schema().peer_init_score()
@@ -22,24 +22,11 @@ fn test_new_connected_peer() {
 }
 
 #[test]
-fn test_add_discovered_address() {
+fn test_add_discovered_addr() {
     let mut peer_store: Box<dyn PeerStore> = Box::new(SqlitePeerStore::temp());
     let peer_id = random_peer_id();
-    peer_store
-        .add_discovered_address(&peer_id, "/ip4/127.0.0.1".to_multiaddr().unwrap())
-        .expect("add discovered address");
+    peer_store.add_discovered_addr(&peer_id, "/ip4/127.0.0.1".to_multiaddr().unwrap());
     assert_eq!(peer_store.peer_addrs(&peer_id, 2).unwrap().len(), 1);
-    peer_store
-        .add_discovered_addresses(
-            &peer_id,
-            vec![
-                "/ip4/127.0.0.1".to_multiaddr().unwrap(),
-                "/ip4/192.168.2.2".to_multiaddr().unwrap(),
-                "/ip4/192.168.2.3".to_multiaddr().unwrap(),
-            ],
-        )
-        .expect("add discovered address");
-    assert_eq!(peer_store.peer_addrs(&peer_id, 4).unwrap().len(), 3);
 }
 
 #[test]
@@ -59,7 +46,7 @@ fn test_update_status() {
     peer_store.update_status(&peer_id, Status::Connected);
     assert_eq!(peer_store.peer_status(&peer_id), Status::Unknown);
     let addr = "/ip4/127.0.0.1".to_multiaddr().unwrap();
-    peer_store.new_connected_peer(&peer_id, addr, SessionType::Server);
+    peer_store.add_connected_peer(&peer_id, addr, SessionType::Server);
     peer_store.update_status(&peer_id, Status::Connected);
     assert_eq!(peer_store.peer_status(&peer_id), Status::Connected);
 }
@@ -71,7 +58,7 @@ fn test_ban_peer() {
     peer_store.ban_peer(&peer_id, Duration::from_secs(10));
     assert!(!peer_store.is_banned(&peer_id));
     let addr = "/ip4/127.0.0.1".to_multiaddr().unwrap();
-    peer_store.new_connected_peer(&peer_id, addr, SessionType::Server);
+    peer_store.add_connected_peer(&peer_id, addr, SessionType::Server);
     peer_store.ban_peer(&peer_id, Duration::from_secs(10));
     assert!(peer_store.is_banned(&peer_id));
 }
@@ -85,9 +72,7 @@ fn test_bootnodes() {
     peer_store.add_bootnode(peer_id.clone(), addr.clone());
     assert_eq!(peer_store.bootnodes(2).len(), 1);
     let peer_id2 = random_peer_id();
-    peer_store
-        .add_discovered_address(&peer_id2, addr.clone())
-        .expect("add discovered address");
+    peer_store.add_discovered_addr(&peer_id2, addr.clone());
     assert_eq!(
         peer_store.bootnodes(3),
         vec![(peer_id2, addr.clone()), (peer_id, addr)]
@@ -103,9 +88,7 @@ fn test_peers_to_attempt() {
     peer_store.add_bootnode(peer_id.clone(), addr.clone());
     assert!(peer_store.peers_to_attempt(1).is_empty());
     let peer_id2 = random_peer_id();
-    peer_store
-        .add_discovered_address(&peer_id2, addr.clone())
-        .expect("add discovered address");
+    peer_store.add_discovered_addr(&peer_id2, addr.clone());
     assert_eq!(peer_store.peers_to_attempt(2).len(), 1);
     peer_store.update_status(&peer_id2, Status::Connected);
     assert!(peer_store.peers_to_attempt(1).is_empty());
@@ -120,9 +103,7 @@ fn test_random_peers() {
     peer_store.add_bootnode(peer_id.clone(), addr.clone());
     assert!(peer_store.random_peers(1).is_empty());
     let peer_id2 = random_peer_id();
-    peer_store
-        .add_discovered_address(&peer_id2, addr.clone())
-        .expect("add discovered address");
+    peer_store.add_discovered_addr(&peer_id2, addr.clone());
     assert_eq!(peer_store.random_peers(2).len(), 1);
     peer_store.update_status(&peer_id2, Status::Connected);
     assert_eq!(peer_store.random_peers(1).len(), 1);
@@ -134,7 +115,7 @@ fn test_delete_peer_info() {
     let addr1 = "/ip4/127.0.0.1".to_multiaddr().unwrap();
     let addr2 = "/ip4/192.163.1.1".to_multiaddr().unwrap();
     for _ in 0..(PEER_STORE_LIMIT - 2) {
-        peer_store.new_connected_peer(&random_peer_id(), addr1.clone(), SessionType::Server);
+        peer_store.add_connected_peer(&random_peer_id(), addr1.clone(), SessionType::Server);
     }
     let evict_target = random_peer_id();
     let fake_target = random_peer_id();
@@ -145,8 +126,8 @@ fn test_delete_peer_info() {
         let faketime_file = faketime::millis_tempfile(recent_not_seen_time.as_secs() * 1000)
             .expect("create faketime file");
         faketime::enable(&faketime_file);
-        peer_store.new_connected_peer(&evict_target, addr1.clone(), SessionType::Server);
-        peer_store.new_connected_peer(&fake_target, addr2, SessionType::Server);
+        peer_store.add_connected_peer(&evict_target, addr1.clone(), SessionType::Server);
+        peer_store.add_connected_peer(&fake_target, addr2, SessionType::Server);
     }
     peer_store.report(&evict_target, Behaviour::FailedToPing);
     peer_store.report(&fake_target, Behaviour::FailedToPing);
@@ -157,7 +138,7 @@ fn test_delete_peer_info() {
             < peer_store.scoring_schema().peer_init_score()
     );
     // should evict evict_target and accept this
-    peer_store.new_connected_peer(&random_peer_id(), addr1, SessionType::Server);
+    peer_store.add_connected_peer(&random_peer_id(), addr1, SessionType::Server);
     // evict_target is evicted in previous step
     assert_eq!(
         peer_store.peer_score_or_default(&evict_target),
